melhorando projeto

+++ CrudProjetoClassLibrary/Entidades/Produto.cs
@@ -10,6 +10,6 @@ namespace CrudProjetoClassLibrary.Entidades
     {
         public int Id { get; set; }
         public string Nome { get; set; }
-        public decimal Preco { get; set; }
+        public float Preco { get; set; }
     }
 }
